Remove testcases from tests.py, tests_with_setup_base.py , tests_with_setup.py and distribute them in other suites. Sanity tests will be tagged separately
