+++ scripts/tests_with_setup_base.py
@@ -57,119 +57,4 @@ class TestSanityBase(testtools.TestCase, ResourcedTestCase, fixtures.TestWithFix
         pass
     # end runTest
 
-    @preposttest_wrapper
-    def test_vn_add_delete(self):
-        '''Test to validate VN creation and deletion.
-            1. Create VN with subnet
-            2. Verify VN against control-node, collector and API
-            3. Delete VN and verify
-        Pass criteria: Step 2 and 3 should pass 
-        '''
-        vn_fixture = self.useFixture(
-            VNFixture(
-                project_name=self.inputs.project_name, connections=self.connections,
-                vn_name='vnxx', inputs=self.inputs, subnets=['22.1.1.0/24']))
-        assert vn_fixture.verify_on_setup()
-        return True
-    # end
-
-    @preposttest_wrapper
-    def test_vm_add_delete(self):
-        ''' Test to validate that a VM creation and deletion passes.
-            1. Create a VN and launch VM within it
-            2. Verify VN and VM against control-node, collector and API
-            3. Delete VM & VN and verify
-        Pass criteria: Step 2 and 3 should pass 
-        '''
-        vm1_name = 'vm_mine'
-        vn_name = 'vn222'
-        vn_subnets = ['11.1.1.0/24']
-        vn_fixture = self.useFixture(
-            VNFixture(
-                project_name=self.inputs.project_name, connections=self.connections,
-                vn_name=vn_name, inputs=self.inputs, subnets=vn_subnets))
-        assert vn_fixture.verify_on_setup()
-        vn_obj = vn_fixture.obj
-        vm1_fixture = self.useFixture(VMFixture(connections=self.connections,
-                                                vn_obj=vn_obj, vm_name=vm1_name, project_name=self.inputs.project_name, image_name='ubuntu'))
-        assert vm1_fixture.verify_on_setup()
-        assert vm1_fixture.wait_till_vm_is_up()
-        return True
-    # end test_vm_add_delete
-
-    @preposttest_wrapper
-    def test_floating_ip(self):
-        '''Test to validate floating-ip Assignment to a VM. 
-            1. Pick VN from resource pool which has VM'in it 
-            2. Create FIP pool for resource FIP VN fvn
-            3. Associate FIP from pool to test VM and verify
-            4. Ping to FIP from test VM
-        Pass criteria: Step 2,3 and 4 should pass
-        '''
-        result = True
-        fip_pool_name = 'some-pool1'
-        fvn_name = self.res.fip_vn_name
-        fvn_fixture = self.res.get_fvn_fixture()
-        vn1_fixture = self.res.get_vn1_fixture()
-        vn1_vm1_fixture = self.res.get_vn1_vm1_fixture()
-        assert vn1_vm1_fixture.verify_on_setup(force=True)
-        assert vn1_vm1_fixture.wait_till_vm_is_up()
-        fvn_vm1_fixture = self.res.get_fvn_vm1_fixture()
-        assert fvn_vm1_fixture.wait_till_vm_is_up()
-        fvn_subnets = self.res.fip_vn_subnets
-        vm1_name = self.res.vn1_vm1_name
-        vn1_name = self.res.vn1_name
-        vn1_subnets = self.res.vn1_subnets
-
-        fip_fixture = self.useFixture(
-            FloatingIPFixture(
-                project_name=self.inputs.project_name, inputs=self.inputs,
-                connections=self.connections, pool_name=fip_pool_name, vn_id=fvn_fixture.vn_id))
-        assert fip_fixture.verify_on_setup()
-        fip_id = fip_fixture.create_and_assoc_fip(
-            fvn_fixture.vn_id, vn1_vm1_fixture.vm_id)
-        assert fip_fixture.verify_fip(fip_id, vn1_vm1_fixture, fvn_fixture)
-        if not vn1_vm1_fixture.ping_with_certainty(fvn_vm1_fixture.vm_ip):
-            result = result and False
-        fip_fixture.disassoc_and_delete_fip(fip_id)
-        if not result:
-            self.logger.error('Test to ping between VMs %s and %s' %
-                              (vn1_vm1_name, fvn_vm1_name))
-            self.logger.info('Checking if the required verifications',
-                             'on the fixtures pass')
-            assert fvn_fixture.verify_on_setup()
-            assert vn1_fixture.verify_on_setup()
-            assert vn1_vm1_fixture.verify_on_setup()
-            assert fvn_vm1_fixture.verify_on_setup()
-            assert result
-        return True
-    # end test_floating_ip
-
-    @preposttest_wrapper
-    def test_ping_within_vn(self):
-        ''' Validate Ping between two VMs within a VN.
-            1. Pick VN from resource pool which has 2 VM's within it 
-            2. Verify VN & VM against control-node, collector and API
-            3. Ping from one VM to another which are launched in same network
-        Pass criteria: Step 2 and 3 should pass
-        '''
-        vn1_name = self.res.vn1_name
-        vn1_subnets = self.res.vn1_subnets
-        vn1_vm1_name = self.res.vn1_vm1_name
-        vn1_vm2_name = self.res.vn1_vm2_name
-        vn1_fixture = self.res.get_vn1_fixture()
-        vm1_fixture = self.res.get_vn1_vm1_fixture()
-        vm1_fixture.wait_till_vm_is_up()
-
-        vm2_fixture = self.res.get_vn1_vm2_fixture()
-        vm2_fixture.wait_till_vm_is_up()
-        if not vm1_fixture.ping_to_ip(vm2_fixture.vm_ip) or \
-           not vm2_fixture.ping_to_ip(vm1_fixture.vm_ip):
-            self.logger.error('Ping between VMs failed,',
-                 ' Verifying the fixtures...')
-            assert vn1_fixture.verify_on_setup()
-            assert vm1_fixture.verify_on_setup()
-            assert vm2_fixture.verify_on_setup()
-        return True
-    # end test_ping_within_vn
 # end TestSanityBase
