+++ scripts/vm_regression/test_vm.py
@@ -14,92 +14,112 @@ from base import BaseVnVmTest
 from common import isolated_creds
 import inspect
 
+import test
+
 class TestBasicVMVN0(BaseVnVmTest):
 
     @classmethod
     def setUpClass(cls):
         super(TestBasicVMVN0, cls).setUpClass()
 
-    def runTest(self):
-        pass
-    #end runTes 
     @preposttest_wrapper
     def test_bring_up_vm_with_control_node_down(self):
         ''' Create VM when there is not active control node. Verify VM comes up fine when all control nodes are back
-        
+
         '''
-        if len(set(self.inputs.bgp_ips)) < 2 :
-            raise self.skipTest("Skiping Test. At least 2 control node required to run the test")
+        if len(set(self.inputs.bgp_ips)) < 2:
+            raise self.skipTest(
+                "Skiping Test. At least 2 control node required to run the test")
         result = True
-        vn1_name='vn30'
-        vn1_subnets=['30.1.1.0/24']
+        vn1_name = 'vn30'
+        vn1_subnets = ['30.1.1.0/24']
 
         # Collecting all the control node details
-        controller_list= []
+        controller_list = []
         for entry in self.inputs.compute_ips:
-            inspect_h= self.agent_inspect[entry]
-            agent_xmpp_status= inspect_h.get_vna_xmpp_connection_status()
+            inspect_h = self.agent_inspect[entry]
+            agent_xmpp_status = inspect_h.get_vna_xmpp_connection_status()
             for entry in agent_xmpp_status:
                 controller_list.append(entry['controller_ip'])
         controller_list = set(controller_list)
 
         # Stop all the control node
         for entry in controller_list:
-            self.logger.info('Stoping the Control service in  %s' %(entry))
-#            self.inputs.stop_service('contrail-control',[entry])
-            self.addCleanup( self.inputs.start_service, 'contrail-control', [entry] )
+            self.logger.info('Stoping the Control service in  %s' % (entry))
+            self.inputs.stop_service('contrail-control', [entry])
+            self.addCleanup(self.inputs.start_service,
+                            'contrail-control', [entry])
         sleep(30)
 
-        vn1_vm1_name= 'vm1'
-        vn1_vm2_name= 'vm2'
-        vn1_fixture= self.useFixture(VNFixture(project_name= self.project.project_name, connections= self.connections,
-                     vn_name=vn1_name, inputs= self.inputs, subnets= vn1_subnets))
-        vm1_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
-                vn_obj= vn1_fixture.obj, vm_name= vn1_vm1_name))
+        vn1_vm1_name = 'vm1'
+        vn1_vm2_name = 'vm2'
+        vn1_fixture = self.useFixture(
+            VNFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_name=vn1_name, inputs=self.inputs, subnets=vn1_subnets))
+        vm1_fixture = self.useFixture(
+            VMFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_obj=vn1_fixture.obj, vm_name=vn1_vm1_name))
 
-        vm2_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
-                vn_obj= vn1_fixture.obj, vm_name= vn1_vm2_name))
+        vm2_fixture = self.useFixture(
+            VMFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_obj=vn1_fixture.obj, vm_name=vn1_vm2_name))
 
         vm1_fixture.verify_vm_launched()
         vm2_fixture.verify_vm_launched()
-        vm1_node_ip=self.inputs.host_data[self.nova_fixture.get_nova_host_of_vm(vm1_fixture.vm_obj)]['host_ip']
-        vm2_node_ip=self.inputs.host_data[self.nova_fixture.get_nova_host_of_vm(vm2_fixture.vm_obj)]['host_ip']
-        inspect_h1= self.agent_inspect[vm1_node_ip]
-        inspect_h2= self.agent_inspect[vm2_node_ip]
-        self.logger.info('Checking TAP interface is created for all VM and  should be in Error state')
+        vm1_node_ip = self.inputs.host_data[
+            self.nova_fixture.get_nova_host_of_vm(vm1_fixture.vm_obj)]['host_ip']
+        vm2_node_ip = self.inputs.host_data[
+            self.nova_fixture.get_nova_host_of_vm(vm2_fixture.vm_obj)]['host_ip']
+        inspect_h1 = self.agent_inspect[vm1_node_ip]
+        inspect_h2 = self.agent_inspect[vm2_node_ip]
+        self.logger.info(
+            'Checking TAP interface is created for all VM and  should be in Error state')
         vm1_tap_intf = None
         vm2_tap_intf = None
-        vm1_tap_intf=inspect_h1.get_vna_tap_interface_by_ip(vm1_fixture.vm_ip)
+        vm1_tap_intf = inspect_h1.get_vna_tap_interface_by_ip(
+            vm1_fixture.vm_ip)
         if vm1_tap_intf is []:
-            self.logger.error('TAP interface is not created for VM %s' %(vn1_vm1_name))
-            result= result and False
+            self.logger.error('TAP interface is not created for VM %s' %
+                              (vn1_vm1_name))
+            result = result and False
         else:
             if vm1_tap_intf[0]['vrf_name'] != '--ERROR--':
-                self.logger.error('TAP interface VRF info should be Error . But currently in %s' %(vm1_tap_intf[0]['vrf_name']))
-                result= result and False
+                self.logger.error(
+                    'TAP interface VRF info should be Error . But currently in %s' %
+                    (vm1_tap_intf[0]['vrf_name']))
+                result = result and False
 
-        vm2_tap_intf=inspect_h2.get_vna_tap_interface_by_ip(vm2_fixture.vm_ip)
+        vm2_tap_intf = inspect_h2.get_vna_tap_interface_by_ip(
+            vm2_fixture.vm_ip)
         if vm2_tap_intf is []:
-                self.logger.error('TAP interface is not created for VM %s' %(vn1_vm2_name))
-                result= result and False
+            self.logger.error('TAP interface is not created for VM %s' %
+                              (vn1_vm2_name))
+            result = result and False
         else:
             if vm2_tap_intf[0]['vrf_name'] != '--ERROR--':
-                self.logger.error('TAP interface VRF info should be Error . But currently in %s' %(vm2_tap_intf[0]['vrf_name']))
-                result= result and False
+                self.logger.error(
+                    'TAP interface VRF info should be Error . But currently in %s' %
+                    (vm2_tap_intf[0]['vrf_name']))
+                result = result and False
 
         self.logger.info('Waiting for 120 sec for cleanup to begin')
-        sleep (120)
+        sleep(120)
         # Check agent should not have any VN info
         for entry in self.inputs.compute_ips:
-            inspect_h= self.agent_inspect[entry]
-            self.logger.info('Checking VN info in agent %s.' %(entry))
+            inspect_h = self.agent_inspect[entry]
+            self.logger.info('Checking VN info in agent %s.' % (entry))
             if inspect_h.get_vna_vn_list()['VNs'] != []:
-                self.logger.error('Agent should not have any VN info present when control node is down')
-                result= result and False
+                self.logger.error(
+                    'Agent should not have any VN info present when control node is down')
+                result = result and False
+
         # Start all the control node
         for entry in controller_list:
-            self.logger.info('Starting the Control service in  %s' %(entry))
-            self.inputs.start_service('contrail-control',[entry])
+            self.logger.info('Starting the Control service in  %s' % (entry))
+            self.inputs.start_service('contrail-control', [entry])
         sleep(10)
 
         self.logger.info('Checking the VM came up properly or not')
@@ -108,11 +128,13 @@ class TestBasicVMVN0(BaseVnVmTest):
         assert vm1_fixture.verify_on_setup()
 
         # Check ping between VM
-        assert vm2_fixture.ping_to_ip( vm1_fixture.vm_ip )
-        if not result :
-            self.logger.error('Test to verify cleanup of agent after control nodes stop Failed')
+        assert vm2_fixture.ping_to_ip(vm1_fixture.vm_ip)
+        if not result:
+            self.logger.error(
+                'Test to verify cleanup of agent after control nodes stop Failed')
             assert result
         return True
+    # end test
 
     @preposttest_wrapper
     def test_broadcast_udp_w_chksum(self):
@@ -261,56 +283,6 @@ class TestBasicVMVN0(BaseVnVmTest):
 
         return True
     #end test_bulk_add_delete
-
-    @preposttest_wrapper
-    def test_diff_proj_same_vn_vm_add_delete(self):
-        ''' Test to validate that a VN and VM with the same name and same subnet can be created in two different projects
-        '''
-        vm_name='vm_mine'
-        vn_name='vn222'
-        ts = time.time()
-        vn_name = '%s_%s'%(inspect.stack()[0][3],str(ts))
-        vn_subnets=['11.1.1.0/24']
-        projects=['project1111', 'project2222']
-        user_list = [('gudi', 'gudi123', 'admin'), ('mal', 'mal123', 'admin')]
-
-        project_fixture1 = self.useFixture(ProjectFixture(project_name = projects[0],vnc_lib_h= self.vnc_lib,username=user_list[0][0],
-            password= user_list[0][1],connections= self.connections))
-        project_inputs1= self.useFixture(ContrailTestInit(self.ini_file, stack_user=project_fixture1.username,
-            stack_password=project_fixture1.password,project_fq_name=['default-domain',projects[0]],logger = self.logger))
-        project_connections1= ContrailConnections(project_inputs1,self.logger)
-
-        project_fixture2 = self.useFixture(ProjectFixture(project_name = projects[1],vnc_lib_h= self.vnc_lib,username=user_list[1][0],
-            password= user_list[1][1],connections= self.connections))
-        project_inputs2= self.useFixture(ContrailTestInit(self.ini_file, stack_user=project_fixture2.username,
-            stack_password=project_fixture2.password,project_fq_name=['default-domain',projects[1]],logger = self.logger))
-        project_connections2= ContrailConnections(project_inputs2,self.logger)
-
-        vn1_fixture= self.useFixture(VNFixture(project_name= projects[0], connections= project_connections1,
-                         vn_name=vn_name, inputs= project_inputs1, subnets= vn_subnets))
-
-        assert vn1_fixture.verify_on_setup()
-        vn1_obj= vn1_fixture.obj
-
-        vn2_fixture= self.useFixture(VNFixture(project_name= projects[1], connections= project_connections2,
-                         vn_name=vn_name, inputs= project_inputs2, subnets= vn_subnets))
-
-        assert vn2_fixture.verify_on_setup()
-        vn2_obj= vn2_fixture.obj
-
-        vm1_fixture= self.useFixture(VMFixture(connections= project_connections1,
-                   vn_obj=vn1_obj, vm_name= vm_name, project_name= projects[0]))
-        vm2_fixture= self.useFixture(VMFixture(connections= project_connections2,
-                    vn_obj=vn2_obj, vm_name= vm_name, project_name= projects[1]))
-        assert vm1_fixture.verify_on_setup()
-        assert vm2_fixture.verify_on_setup()
-        if not vm1_fixture.agent_label == vm2_fixture.agent_label :
-            self.logger.info("Correct label assigment")
-        else :
-            self.logger.error("The same label has been assigned for both the VMs")
-            return False
-        return True
-    #end test_diff_proj_same_vn_vm_add_delete
 
     @preposttest_wrapper
     def test_disassociate_vn_from_vm(self):
@@ -676,9 +648,6 @@ echo "Hello World.  The time is now $(date -R)!" | tee /tmp/output.txt
         self.nova_fixture.wait_till_vm_is_up( vm4_fixture.vm_obj )
         assert not vm4_fixture.ping_to_ip( vm3_fixture.vm_ip )
         return True
-#class TestBasicVMVN0XML(TestBasicVMVN0):
-#    _interface = 'xml'
-#    pass
 
     @preposttest_wrapper
     def test_diff_proj_same_vn_vm_add_delete(self):
@@ -999,6 +968,88 @@ class TestBasicVMVN2(BaseVnVmTest):
     #end test_ping_on_broadcast_multicast_with_frag
 
     @preposttest_wrapper
+    def test_ping_on_broadcast_multicast(self):
+        ''' Validate Ping on subnet broadcast,link local multucast,network broadcast .
+
+        '''
+        result = True
+        vn1_name = get_random_name('vn1')
+        vn1_subnets = ["192.168.1.0/24"]
+        ping_count = '5'
+        vn1_vm1_name = get_random_name('vn1_vm1')
+        vn1_vm2_name = get_random_name('vn1_vm2')
+        vn1_vm3_name = get_random_name('vn1_vm3')
+        vn1_vm4_name = get_random_name('vn1_vm4')
+        vn1_fixture = self.create_vn(vn1_name, vn1_subnets)
+        vm1_fixture = self.create_vm(vn1_fixture, vn1_vm1_name)
+        vm2_fixture = self.create_vm(vn1_fixture, vn1_vm2_name)
+        vm3_fixture = self.create_vm(vn1_fixture, vn1_vm3_name)
+        vm4_fixture = self.create_vm(vn1_fixture, vn1_vm4_name)
+        assert vm1_fixture.wait_till_vm_is_up()
+        assert vm2_fixture.wait_till_vm_is_up()
+        assert vm3_fixture.wait_till_vm_is_up()
+        assert vm4_fixture.wait_till_vm_is_up()
+        # Geting the VM ips
+        vm1_ip = vm1_fixture.vm_ip
+        vm2_ip = vm2_fixture.vm_ip
+        vm3_ip = vm3_fixture.vm_ip
+        vm4_ip = vm4_fixture.vm_ip
+        ip_list = [vm1_ip, vm2_ip, vm3_ip, vm4_ip]
+        bcast_ip = str(IPNetwork(vn1_subnets[0]).broadcast)
+        list_of_ip_to_ping = [bcast_ip, '224.0.0.1', '255.255.255.255']
+        # passing command to vms so that they respond to subnet broadcast
+        cmd = ['echo 0 > /proc/sys/net/ipv4/icmp_echo_ignore_broadcasts']
+        vm_fixtures = [vm1_fixture, vm2_fixture, vm3_fixture, vm4_fixture]
+        for vm in vm_fixtures:
+            print 'Running cmd for %s' % vm.vm_name
+            for i in range(3):
+                try:
+                    self.logger.info("Retry %s" % (i))
+                    ret = vm.run_cmd_on_vm(cmds=cmd, as_sudo=True)
+                    if not ret:
+                        for vn in vm.vn_fq_names:
+                            vm.ping_vm_from_host(vn)
+                        raise Exception
+                except Exception as e:
+                    time.sleep(5)
+                    self.logger.exception("Got exception as %s" % (e))
+                else:
+                    break
+        for dst_ip in list_of_ip_to_ping:
+            self.logger.info('pinging from %s to %s' % (vm1_ip, dst_ip))
+# pinging from Vm1 to subnet broadcast
+            ping_output = vm1_fixture.ping_to_ip(
+                dst_ip, return_output=True, count=ping_count, other_opt='-b')
+            self.logger.info("ping output : \n %s" % (ping_output))
+            expected_result = ' 0% packet loss'
+            if not expected_result in ping_output:
+                self.logger.error('Expected 0% packet loss seen!')
+                self.logger.error('Ping result : %s' % (ping_output))
+                result = result and True
+# getting count of ping response from each vm
+            string_count_dict = {}
+            string_count_dict = get_string_match_count(ip_list, ping_output)
+            self.logger.info("output %s" % (string_count_dict))
+            self.logger.info(
+                "There should be atleast 4 echo reply from each ip")
+            for k in ip_list:
+                # this is a workaround : ping utility exist as soon as it gets
+                # one response
+#                assert (string_count_dict[k] >= (int(ping_count) - 1))
+                if not string_count_dict[k] >= (int(ping_count) - 1):
+                    self.logger.error('Seen %s reply instead of atleast %s' % (
+                        (int(ping_count) - 1)))
+                    result = result and False
+        if not result:
+            self.logger.error('There were errors. Verifying VM fixtures')
+            assert vm1_fixture.verify_on_setup()
+            assert vm2_fixture.verify_on_setup()
+            assert vm3_fixture.verify_on_setup()
+            assert vm4_fixture.verify_on_setup()
+        return True
+    # end subnet ping
+
+    @preposttest_wrapper
     def test_ping_within_vn(self):
         ''' Validate Ping between two VMs within a VN. 
         
@@ -1016,10 +1067,8 @@ class TestBasicVMVN2(BaseVnVmTest):
                 vn_obj= vn1_fixture.obj, vm_name= vn1_vm1_name))
         vm2_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
                 vn_obj= vn1_fixture.obj, vm_name= vn1_vm2_name))
-        assert vm1_fixture.verify_on_setup()
-        assert vm2_fixture.verify_on_setup()
-        self.nova_fixture.wait_till_vm_is_up( vm1_fixture.vm_obj )
-        self.nova_fixture.wait_till_vm_is_up( vm2_fixture.vm_obj )
+        assert vm1_fixture.wait_till_vm_is_up()
+        assert vm2_fixture.wait_till_vm_is_up()
         assert vm1_fixture.ping_to_ip( vm2_fixture.vm_ip )
         assert vm2_fixture.ping_to_ip( vm1_fixture.vm_ip )
         return True
@@ -1037,19 +1086,17 @@ class TestBasicVMVN2(BaseVnVmTest):
         ts = time.time()
         vn1_name = '%s_%s'%(inspect.stack()[0][3],str(ts))
         #vn1_subnets=['30.1.1.0/24']
-        vn1_vm1_name= 'vm1'
-        vn1_vm2_name= 'vm2'
+        vn1_vm1_name = 'vm1'
+        vn1_vm2_name = 'vm2'
         vn1_fixture= self.useFixture(VNFixture(project_name= self.project.project_name, connections= self.connections,
                      vn_name=vn1_name, inputs= self.inputs, subnets= vn1_subnets))
         assert vn1_fixture.verify_on_setup()
-        vm1_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
+        vm1_fixture = self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
                 vn_obj= vn1_fixture.obj, vm_name= vn1_vm1_name))
-        vm2_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
+        vm2_fixture = self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
                 vn_obj= vn1_fixture.obj, vm_name= vn1_vm2_name))
-        assert vm1_fixture.verify_on_setup()
-        assert vm2_fixture.verify_on_setup()
-        self.nova_fixture.wait_till_vm_is_up( vm1_fixture.vm_obj )
-        self.nova_fixture.wait_till_vm_is_up( vm2_fixture.vm_obj )
+        assert vm1_fixture.wait_till_vm_is_up()
+        assert vm2_fixture.wait_till_vm_is_up()
         assert vm1_fixture.ping_to_ip( vm2_fixture.vm_ip )
         assert vm2_fixture.ping_to_ip( vm1_fixture.vm_ip )
         #Geting the VM ips
@@ -1191,114 +1238,131 @@ class TestBasicVMVN2(BaseVnVmTest):
     def test_process_restart_in_policy_between_vns(self):
         ''' Test to validate that with policy having rule to check icmp fwding between VMs on different VNs , ping between VMs should pass
         with process restarts
+            1. Pick 2 VN's from resource pool which has one VM each
+            2. Create policy with icmp allow rule between those VN's and bind it networks
+            3. Ping from one VM to another VM
+            4. Restart process 'vrouter' and 'control' on setup
+            5. Ping again between VM's after process restart
+        Pass criteria: Step 2,3,4 and 5 should pass
         '''
-        result = True ; msg = []
-        vn1_name='vn40'
-        vn1_subnets=['40.1.1.0/24']
-        vn2_name='vn41'
-        vn2_subnets=['41.1.1.0/24']
-        policy1_name= 'policy1'
-        policy2_name= 'policy2'
-        rules= [
-            {  
-               'direction'     : '<>', 'simple_action' : 'pass',
-               'protocol'      : 'icmp',
-               'source_network': vn1_name,
-               'dest_network'  : vn2_name,
-             }, 
-                ]
-        rev_rules= [
+        vn1_name = get_random_name('vn1')
+        vn1_subnets = ["192.168.1.0/24"]
+        vn2_name = get_random_name('vn2')
+        vn2_subnets = ["192.168.2.0/24"]
+        policy1_name = 'policy1'
+        policy2_name = 'policy2'
+        rules = [
             {
-               'direction'     : '<>', 'simple_action' : 'pass',
-               'protocol'      : 'icmp',
-               'source_network': vn2_name,
-               'dest_network'  : vn1_name,
-             },
-                ]
-        policy1_fixture= self.useFixture( PolicyFixture( policy_name= policy1_name, rules_list= rules, inputs= self.inputs,
-                                    connections= self.connections ))
-        policy2_fixture= self.useFixture( PolicyFixture( policy_name= policy2_name, rules_list= rev_rules, inputs= self.inputs,
-                                    connections= self.connections ))
-        vn1_fixture= self.useFixture(VNFixture(project_name= self.project.project_name, connections= self.connections,
-                     vn_name=vn1_name, inputs= self.inputs, subnets= vn1_subnets, policy_objs=[policy1_fixture.policy_obj]))
+                'direction': '<>', 'simple_action': 'pass',
+                'protocol': 'icmp',
+                'source_network': vn1_name,
+                'dest_network': vn2_name,
+            },
+        ]
+        rev_rules = [
+            {  
+                'direction': '<>', 'simple_action': 'pass',
+                'protocol': 'icmp',
+                'source_network': vn2_name,
+                'dest_network': vn1_name,
+            },
+        ]
+        policy1_fixture = self.useFixture(
+            PolicyFixture(
+                policy_name=policy1_name, rules_list=rules, inputs=self.inputs,
+                connections=self.connections))
+        policy2_fixture = self.useFixture(
+            PolicyFixture(
+                policy_name=policy2_name, 
+                rules_list=rev_rules, inputs=self.inputs,
+                connections=self.connections))
+        vn1_fixture = self.create_vn(vn1_name, vn1_subnets)
         assert vn1_fixture.verify_on_setup()
-        vn2_fixture= self.useFixture(VNFixture(project_name= self.project.project_name, connections= self.connections,
-                     vn_name=vn2_name, inputs= self.inputs, subnets= vn2_subnets, policy_objs=[policy2_fixture.policy_obj]))
+        vn1_fixture.bind_policies(
+            [policy1_fixture.policy_fq_name], vn1_fixture.vn_id)
+        self.addCleanup(vn1_fixture.unbind_policies,
+                        vn1_fixture.vn_id, [policy1_fixture.policy_fq_name])
+        vn2_fixture = self.create_vn(vn2_name, vn2_subnets)
         assert vn2_fixture.verify_on_setup()
-        vn1_vm1_name= 'vm1'
-        vn1_vm2_name= 'vm2'
-        vm1_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
-                vn_obj= vn1_fixture.obj, vm_name= vn1_vm1_name))
-        vm2_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
-                vn_obj= vn2_fixture.obj, vm_name= vn1_vm2_name))
-        assert vm1_fixture.verify_on_setup()
-        assert vm2_fixture.verify_on_setup()
-        self.nova_fixture.wait_till_vm_is_up( vm1_fixture.vm_obj )
-        self.nova_fixture.wait_till_vm_is_up( vm2_fixture.vm_obj )
-        self.logger.info ("Verify ping to vm %s" %(vn1_vm2_name))
-        ret= vm1_fixture.ping_with_certainty( vm2_fixture.vm_ip, expectation=True )
-        result_msg= "vm ping test result to vm %s is: %s" %(vn1_vm2_name, ret)
-        self.logger.info (result_msg)
-        if ret != True : result= False; msg.extend([result_msg, policy1_name])
-        self.assertEqual(result, True, msg)
+        vn2_fixture.bind_policies(
+            [policy2_fixture.policy_fq_name], vn2_fixture.vn_id)
+        self.addCleanup(vn2_fixture.unbind_policies,
+                        vn2_fixture.vn_id, [policy2_fixture.policy_fq_name])
+        vn1_vm1_name = get_random_name('vn1_vm1')
+        vn2_vm1_name = get_random_name('vn2_vm1')
+        vm1_fixture = self.create_vm(vn1_fixture, vn1_vm1_name)
+        vm2_fixture = self.create_vm(vn2_fixture, vn2_vm1_name)
+        assert vm1_fixture.wait_till_vm_is_up()
+        assert vm2_fixture.wait_till_vm_is_up()
+        assert vm1_fixture.ping_with_certainty(vm2_fixture.vm_ip)
 
         for compute_ip in self.inputs.compute_ips:
-               pass
-#                self.inputs.restart_service('contrail-vrouter',[compute_ip])
+            self.inputs.restart_service('contrail-vrouter', [compute_ip])
         for bgp_ip in self.inputs.bgp_ips:
-               pass
-#            self.inputs.restart_service('contrail-control',[bgp_ip])
-        sleep(30)
-        self.logger.info('Sleeping for 30 seconds')
-        vn1_vm3_name= 'vm3'
-        vn1_vm4_name= 'vm4'
-        vm3_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
-                    vn_obj= vn1_fixture.obj, vm_name= vn1_vm3_name))
-        vm4_fixture= self.useFixture(VMFixture(project_name= self.project.project_name, connections= self.connections,
-                    vn_obj= vn2_fixture.obj, vm_name= vn1_vm4_name))
+            self.inputs.restart_service('contrail-control', [bgp_ip])
+        self.logger.info('Sleeping for 10 seconds')
+        sleep(10)
+        vn1_vm2_name = get_random_name('vn1_vm2')
+        vn2_vm2_name = get_random_name('vn2_vm2')
+        vm3_fixture = self.create_vm(vn1_fixture, vn1_vm2_name)
         assert vm3_fixture.verify_on_setup()
+        vm4_fixture = self.create_vm(vn2_fixture, vn2_vm2_name)
         assert vm4_fixture.verify_on_setup()
-        self.nova_fixture.wait_till_vm_is_up( vm3_fixture.vm_obj )
-        self.nova_fixture.wait_till_vm_is_up( vm4_fixture.vm_obj )
-        self.logger.info ("Verify ping to vm %s" %(vn1_vm4_name))
-        ret= vm3_fixture.ping_with_certainty( vm4_fixture.vm_ip, expectation=True )
-        result_msg= "vm ping test result to vm %s is: %s" %(vn1_vm4_name, ret)
-        self.logger.info (result_msg)
-        if ret != True : result= False; msg.extend([result_msg, policy1_name])
-        self.assertEqual(result, True, msg)
-        return True
-     #end test_process_restart_in_policy_between_vns
+        vm3_fixture.wait_till_vm_is_up()
+        vm4_fixture.wait_till_vm_is_up()
+        assert vm3_fixture.ping_with_certainty(vm4_fixture.vm_ip)
+
+# end test_process_restart_in_policy_between_vns
 
     @preposttest_wrapper
     def test_process_restart_with_multiple_vn_vm(self):
-        ''' Test to validate that multiple VM creation and deletion passes.
         '''
-        vm1_name='vm_mine'
-        vn_name='vn222'
-        ts = time.time()
-        vn_name = '%s_%s'%(inspect.stack()[0][3],str(ts))
-        vn_subnets=['11.1.1.0/24']
-        vn_count_for_test=32
+         Description:Test to validate that multiple VM creation and deletion passes.
+
+           1.Create 32 vn and 1 vm in each vn
+           2.Restart vrouter service in each compute node
+           3.Verify all vns /vms are fine after restart - fails otherwise
+
+          Maintainer: sandipd@juniper.net
+        '''
+        vm1_name = 'vm_mine'
+        vn_name = 'vn222'
+        vn_subnets = ['11.1.1.0/24']
+        vn_count_for_test = 32
         if (len(self.inputs.compute_ips) == 1):
-            vn_count_for_test=10
-        vm_fixture= self.useFixture(create_multiple_vn_and_multiple_vm_fixture (connections= self.connections,
-                     vn_name=vn_name, vm_name=vm1_name, inputs= self.inputs , project_name= self.project.project_name,
-                      subnets= vn_subnets,vn_count=vn_count_for_test,vm_count=1,subnet_count=1,image_name='ubuntu'))
-        time.sleep(100)
+            vn_count_for_test = 2
+        try:
+            vm_fixture = self.useFixture(
+                create_multiple_vn_and_multiple_vm_fixture(
+                    connections=self.connections,
+                    vn_name=vn_name, vm_name=vm1_name, inputs=self.inputs, project_name=self.inputs.project_name,
+                    subnets=vn_subnets, vn_count=vn_count_for_test, vm_count=1, subnet_count=1,
+                    image_name='cirros-0.3.0-x86_64-uec', flavor='m1.tiny'))
+            compute_ip = []
+            time.sleep(100)
+        except Exception as e:
+            self.logger.exception("Got exception as %s" % (e))
+
         try:
             assert vm_fixture.verify_vms_on_setup()
+            assert vm_fixture.wait_till_vms_are_up()
             assert vm_fixture.verify_vns_on_setup()
         except Exception as e:
-            self.logger.exception("Got exception as %s"%(e))
-        compute_ip=[]
+            self.logger.exception("Got exception as %s" % (e))
+
         for vmobj in vm_fixture.vm_obj_dict.values():
-            vm_host_ip=vmobj.vm_node_ip
+            vm_host_ip = vmobj.vm_node_ip
             if vm_host_ip not in compute_ip:
                 compute_ip.append(vm_host_ip)
-#        self.inputs.restart_service('contrail-vrouter',compute_ip)
-        sleep(50)
-        for vmobj in vm_fixture.vm_obj_dict.values():
-            assert vmobj.verify_on_setup()
+        self.inputs.restart_service('contrail-vrouter', compute_ip)
+        sleep(30)
+        try:
+            assert vm_fixture.verify_vms_on_setup()
+            assert vm_fixture.wait_till_vms_are_up()
+#            for vmobj in vm_fixture.vm_obj_dict.values():
+#                assert vmobj.verify_on_setup()
+        except Exception as e:
+            self.logger.exception("got exception as %s" % (e))
         return True
 
     @preposttest_wrapper
@@ -2638,6 +2702,398 @@ class TestBasicVMVN6(BaseVnVmTest):
             self.logger.info('Sleeping for 20s')
             sleep(10)
         self.assertEqual(result, True, msg)
+
+    @preposttest_wrapper
+    def test_control_node_switchover(self):
+        ''' Stop the control node and check peering with agent fallback to other control node.
+            1. Pick one VN from respource pool which has 2 VM's in it
+            2. Verify ping between VM's
+            3. Find active control node in cluster by agent inspect
+            4. Stop control service on active control node
+            5. Verify agents are connected to new active control-node using xmpp connections
+            6. Bring back control service on previous active node
+            7. Verify ping between VM's again after bringing up control serveice
+        Pass criteria: Step 2,5 and 7 should pass
+        '''
+        if len(set(self.inputs.bgp_ips)) < 2:
+            self.logger.info(
+                "Skiping Test. At least 2 control node required to run the test")
+            raise self.skipTest(
+                "Skiping Test. At least 2 control node required to run the test")
+        result = True
+        vn1_name = get_random_name('vn1')
+        vn1_subnets = ['192.168.1.0/24']
+        vn1_vm1_name = get_random_name('vn1_vm1')
+        vn1_vm2_name = get_random_name('vn1_vm2')
+        vn1_fixture = self.create_vn(vn1_name, vn1_subnets)
+        assert vn1_fixture.verify_on_setup()
+        vm1_fixture = self.create_vm(vn1_fixture, vn1_vm1_name)
+        assert vm1_fixture.wait_till_vm_is_up()
+        vm2_fixture = self.create_vm(vn1_fixture, vn1_vm2_name)
+        assert vm2_fixture.wait_till_vm_is_up()
+        assert vm1_fixture.ping_to_ip(vm2_fixture.vm_ip)
+        assert vm2_fixture.ping_to_ip(vm1_fixture.vm_ip)
+
+        # Figuring the active control node
+        active_controller = None
+        self.agent_inspect = self.connections.agent_inspect
+        inspect_h = self.agent_inspect[vm1_fixture.vm_node_ip]
+        agent_xmpp_status = inspect_h.get_vna_xmpp_connection_status()
+        for entry in agent_xmpp_status:
+            if entry['cfg_controller'] == 'Yes':
+                active_controller = entry['controller_ip']
+        active_controller_host_ip = self.inputs.host_data[
+            active_controller]['host_ip']
+        self.logger.info('Active control node from the Agent %s is %s' %
+                         (vm1_fixture.vm_node_ip, active_controller_host_ip))
+
+        # Stop on Active node
+        self.logger.info('Stoping the Control service in  %s' %
+                         (active_controller_host_ip))
+        self.inputs.stop_service(
+            'contrail-control', [active_controller_host_ip])
+        sleep(5)
+
+        # Check the control node shifted to other control node
+        new_active_controller = None
+        new_active_controller_state = None
+        inspect_h = self.agent_inspect[vm1_fixture.vm_node_ip]
+        agent_xmpp_status = inspect_h.get_vna_xmpp_connection_status()
+        for entry in agent_xmpp_status:
+            if entry['cfg_controller'] == 'Yes':
+                new_active_controller = entry['controller_ip']
+                new_active_controller_state = entry['state']
+        new_active_controller_host_ip = self.inputs.host_data[
+            new_active_controller]['host_ip']
+        self.logger.info('Active control node from the Agent %s is %s' %
+                         (vm1_fixture.vm_node_ip, new_active_controller_host_ip))
+        if new_active_controller_host_ip == active_controller_host_ip:
+            self.logger.error(
+                'Control node switchover fail. Old Active controlnode was %s and new active control node is %s' %
+                (active_controller_host_ip, new_active_controller_host_ip))
+            result = False
+
+        if new_active_controller_state != 'Established':
+            self.logger.error(
+                'Agent does not have Established XMPP connection with Active control node')
+            result = result and False
+
+        # Start the control node service again
+        self.logger.info('Starting the Control service in  %s' %
+                         (active_controller_host_ip))
+        self.inputs.start_service(
+            'contrail-control', [active_controller_host_ip])
+
+        # Check the BGP peering status from the currently active control node
+        sleep(5)
+        cn_bgp_entry = self.cn_inspect[
+            new_active_controller_host_ip].get_cn_bgp_neigh_entry()
+        for entry in cn_bgp_entry:
+            if entry['state'] != 'Established':
+                result = result and False
+                self.logger.error(
+                    'With Peer %s peering is not Established. Current State %s ' %
+                    (entry['peer'], entry['state']))
+
+        # Check the ping
+        self.logger.info('Checking the ping between the VM again')
+        assert vm1_fixture.ping_to_ip(vm2_fixture.vm_ip)
+        assert vm2_fixture.ping_to_ip(vm1_fixture.vm_ip)
+
+        if not result:
+            self.logger.error('Switchover of control node failed')
+            assert result
+        return True
+
+    # end test_control_node_switchover
+
+    @preposttest_wrapper
+    def test_agent_cleanup_with_control_node_stop(self):
+        ''' Stop all the control node and verify the cleanup process in agent
+
+        '''
+        raise self.skipTest("Skiping a failing test")
+        if len(set(self.inputs.bgp_ips)) < 2:
+            raise self.skipTest(
+                "Skiping Test. At least 2 control node required to run the test")
+        result = True
+        vn1_name = 'vn30'
+        vn1_subnets = ['30.1.1.0/24']
+        vn1_vm1_name = 'vm1'
+        vn1_vm2_name = 'vm2'
+        vn1_fixture = self.useFixture(
+            VNFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_name=vn1_name, inputs=self.inputs, subnets=vn1_subnets))
+        assert vn1_fixture.verify_on_setup()
+        vm1_fixture = self.useFixture(
+            VMFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_obj=vn1_fixture.obj, vm_name=vn1_vm1_name))
+        vm2_fixture = self.useFixture(
+            VMFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_obj=vn1_fixture.obj, vm_name=vn1_vm2_name))
+        assert vm1_fixture.verify_on_setup()
+        assert vm2_fixture.verify_on_setup()
+        vm1_fixture.wait_till_vm_is_up()
+        vm2_fixture.wait_till_vm_is_up()
+
+        assert vm2_fixture.ping_to_ip(vm1_fixture.vm_ip)
+        # Collecting all the control node details
+        controller_list = []
+        inspect_h = self.agent_inspect[vm1_fixture.vm_node_ip]
+        agent_xmpp_status = inspect_h.get_vna_xmpp_connection_status()
+        for entry in agent_xmpp_status:
+            controller_list.append(entry['controller_ip'])
+        list_of_vm = inspect_h.get_vna_vm_list()
+
+        # Stop all the control node
+        for entry in controller_list:
+            self.logger.info('Stoping the Control service in  %s' % (entry))
+            self.inputs.stop_service('contrail-control', [entry])
+            self.addCleanup(self.inputs.start_service,
+                            'contrail-control', [entry])
+            sleep(5)
+
+        # Wait for cleanup to begin
+        sleep(120)
+        # Verify VM entry is removed from the agent introspect
+        vm_id_list = inspect_h.get_vna_vm_list()
+        if vm1_fixture.vm_id in vm_id_list:
+            result = result and False
+            self.logger.error(
+                'VM %s is still present in Agent Introspect.Cleanup not working when all control node shut' %
+                (vm1_fixture.vm_name))
+        if vm2_fixture.vm_id in vm_id_list:
+            result = result and False
+            self.logger.error(
+                'VM %s is still present in Agent Introspect.Cleanup not working when all control node shut' %
+                (vm2_fixture.vm_name))
+
+        # TODO Verify the IF-Map entry
+        # Start all the control node
+        for entry in controller_list:
+            self.logger.info('Starting the Control service in  %s' % (entry))
+            self.inputs.start_service('contrail-control', [entry])
+            sleep(30)
+        # Check everything came up fine
+        vm_id_list = inspect_h.get_vna_vm_list()
+        if vm1_fixture.vm_id not in vm_id_list or vm2_fixture.vm_id not in vm_id_list:
+            result = result and False
+            self.logger.error(
+                'After starting the service all the VM entry did not came up properly')
+
+        self.logger.info('Checking the VM came up properly or not')
+        assert vm2_fixture.verify_on_setup()
+        assert vm1_fixture.verify_on_setup()
+        if not result:
+            self.logger.error(
+                'Test to verify cleanup of agent after control nodes stop Failed')
+            assert result
+        return True
+    # end test_agent_cleanup_with_control_node_stop
+
+    @preposttest_wrapper
+    def test_metadata_service(self):
+        '''
+          Description: Test to validate metadata service on VM creation.
+
+               1.Verify from global-vrouter-config if metadata configures or not - fails otherwise
+               2.Create a shell script which writes  'hello world ' in a file in /tmp and save the script on the nova api node
+               3.Create a vm with userdata pointing to that script - script should get executed during vm boot up
+               4.Go to the vm and verify if the file with 'hello world ' written saved in /tmp of the vm - fails otherwise
+            Maintainer: sandipd@juniper.net
+        '''
+
+        gvrouter_cfg_obj = self.api_s_inspect.get_global_vrouter_config()
+        ln_svc = gvrouter_cfg_obj.get_link_local_service()
+        if ln_svc:
+            self.logger.info(
+                "Metadata configured in global_vrouter_config as %s" %
+                (str(ln_svc)))
+        else:
+            self.logger.warn(
+                "Metadata NOT configured in global_vrouter_config")
+            result = False
+            assert result
+            return True
+
+        text = """#!/bin/sh
+echo "Hello World.  The time is now $(date -R)!" | tee /tmp/output.txt
+               """
+        try:
+            with open("/tmp/metadata_script.txt", "w") as f:
+                f.write(text)
+        except Exception as e:
+            self.logger.exception(
+                "Got exception while creating /tmp/metadata_script.txt as %s" % (e))
+        vn_name = 'vn2_metadata'
+        vm1_name = 'vm_in_vn2_metadata'
+        vn_subnets = ['11.1.1.0/24']
+        vn_fixture = self.useFixture(
+            VNFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_name=vn_name, inputs=self.inputs, subnets=vn_subnets))
+        assert vn_fixture.verify_on_setup()
+        vn_obj = vn_fixture.obj
+        vm1_fixture = self.useFixture(VMFixture(connections=self.connections,
+                                                vn_obj=vn_obj, vm_name=vm1_name, project_name=self.inputs.project_name,
+                                                image_name='cirros-0.3.0-x86_64-uec', userdata='/tmp/metadata_script.txt',
+                                                flavor='m1.tiny'))
+
+        assert vm1_fixture.verify_on_setup()
+        vm1_fixture.wait_till_vm_is_up()
+
+        cmd = 'ls /tmp/'
+        for i in range(3):
+            try:
+                self.logger.info("Retry %s" % (i))
+                ret = vm1_fixture.run_cmd_on_vm(cmds=[cmd])
+#                if 'Connection refused' in ret:
+                if not ret:
+                    raise Exception
+            except Exception as e:
+                time.sleep(5)
+                self.logger.exception("Got exception as %s" % (e))
+            else:
+                break
+        self.logger.info("ret : %s" % (ret))
+        result = False
+        for elem in ret.values():
+            if 'output.txt' in elem:
+                result = True
+                break
+        if not result:
+            self.logger.warn(
+                "metadata_script.txt did not get executed in the vm")
+        else:
+            self.logger.info("Printing the output.txt :")
+            cmd = 'cat /tmp/output.txt'
+            ret = vm1_fixture.run_cmd_on_vm(cmds=[cmd])
+            self.logger.info("%s" % (ret.values()))
+            for elem in ret.values():
+                if 'Hello World' in elem:
+                    result = True
+                else:
+                    self.logger.warn(
+                        "metadata_script.txt did not get executed in the vm...output.txt does not contain proper output")
+                    result = False
+
+        assert result
+        return True
+
+    @preposttest_wrapper
+    def test_generic_link_local_service(self):
+        '''
+         Description: Test to validate generic linklocal service - running nova list from vm.
+            1.Create generic link local service to be able to wget to jenkins
+            2.Create a vm
+            3.Try wget to jenkins - passes if successful else fails
+
+         Maintainer: sandipd@juniper.net
+        '''
+
+        vn_name = 'vn2_metadata'
+        vm1_name = 'nova_client_vm'
+        vn_subnets = ['11.1.1.0/24']
+        vn_fixture = self.useFixture(
+            VNFixture(
+                project_name=self.inputs.project_name, connections=self.connections,
+                vn_name=vn_name, inputs=self.inputs, subnets=vn_subnets))
+        #assert vn_fixture.verify_on_setup()
+        vn_obj = vn_fixture.obj
+        vm1_fixture = self.useFixture(VMFixture(connections=self.connections,
+                                                vn_obj=vn_obj, vm_name=vm1_name, project_name=self.inputs.project_name,
+                                                image_name='ubuntu-traffic'))
+
+        assert vm1_fixture.verify_on_setup()
+        vm1_fixture.wait_till_vm_is_up()
+
+        metadata_args = "--admin_user admin \
+         --admin_password contrail123 --linklocal_service_name generic_link_local\
+         --linklocal_service_ip 169.254.1.1\
+         --linklocal_service_port 8090\
+         --ipfabric_service_ip %s\
+         --ipfabric_service_port 5000\
+         --oper add" % (self.inputs.openstack_ip)
+
+        if not self.inputs.devstack:
+            cmd = "python /opt/contrail/utils/provision_linklocal.py %s" % (metadata_args)
+        else:
+            cmd = "python /opt/stack/contrail/controller/src/config/utils/provision_linklocal.py %s" % (
+                metadata_args)
+
+        link_local_args = "--admin_user admin \
+         --admin_password contrail123 --linklocal_service_name vim\
+         --linklocal_service_ip 169.254.1.2\
+         --linklocal_service_port 80\
+         --ipfabric_dns_service_name www.vim.org\
+         --ipfabric_service_port 80\
+         --oper add"
+
+        if not self.inputs.devstack:
+            cmd = "python /opt/contrail/utils/provision_linklocal.py %s" % (link_local_args)
+        else:
+            cmd = "python /opt/stack/contrail/controller/src/config/utils/provision_linklocal.py %s" % (
+                link_local_args)
+
+        args = shlex.split(cmd)
+        process = Popen(args, stdout=PIPE)
+        stdout, stderr = process.communicate()
+        if stderr:
+            self.logger.warn(
+                "Linklocal service could not be created, err : \n %s" % (stderr))
+        else:
+            self.logger.info("%s" % (stdout))
+        cmd = 'wget http://169.254.1.2:80'
+
+        for i in range(3):
+            try:
+                self.logger.info("Retry %s" % (i))
+                ret = vm1_fixture.run_cmd_on_vm(cmds=[cmd])
+#                if 'Connection refused' in ret:
+                if not ret:
+                    raise Exception
+            except Exception as e:
+                time.sleep(5)
+                self.logger.exception("Got exception as %s" % (e))
+            else:
+                break
+        if ret:
+            if '200 OK' in str(ret):
+                self.logger.info("Generic metadata worked")
+                result = True
+            if 'Connection timed out' in str(ret):
+                self.logger.warn("Generic metadata did NOT work")
+                result = False
+
+        link_local_args = "--admin_user admin \
+         --admin_password contrail123 --linklocal_service_name vim\
+         --linklocal_service_ip 169.254.1.2\
+         --linklocal_service_port 80\
+         --ipfabric_dns_service_name www.vim.org\
+         --ipfabric_service_port 80\
+         --oper delete"
+
+        if not self.inputs.devstack:
+            cmd = "python /opt/contrail/utils/provision_linklocal.py %s" % (link_local_args)
+        else:
+            cmd = "python /opt/stack/contrail/controller/src/config/utils/provision_linklocal.py %s" % (
+                link_local_args)
+
+        args = shlex.split(cmd)
+        process = Popen(args, stdout=PIPE)
+        stdout, stderr = process.communicate()
+        if stderr:
+            self.logger.warn(
+                "Linklocal service could not be deleted, err : \n %s" % (stderr))
+        else:
+            self.logger.info("%s" % (stdout))
+        assert result
+        return True
+    # end test_generic_link_local_service
+
 
 #class TestBasicVMVN6XML(TestBasicVMVN6):
 #    _interface = 'xml'
